requirements.txt
fastapi
uvicorn
llama-index
qdrant-client
pypdf
cryptography
llama-index-llms-ollama
llama-index-vector-stores-qdrant
llama-index-embeddings-fastembed
llama-index-readers-file
fastembed
python-dotenv
requests
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# OpenTelemetry Core
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0

# OpenTelemetry Exporters
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-requests>=0.43b0

# Semantic Conventions
opentelemetry-semantic-conventions>=0.43b0
